+++ Personal-Testing-System/DTOs/SubsequenceDto.cs
@@ -8,6 +8,6 @@ namespace Personal_Testing_System.DTOs
         public int Id { get; set; }
         public string? Text { get; set; }
         public string? IdQuestion { get; set; }
-        public byte? Number { get; set; }
+        public int? Number { get; set; }
     }
 }
